% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-8.R
\name{baselineNorm}
\alias{baselineNorm}
\title{Signal processing}
\usage{
baselineNorm(X, baseLength = 30)
}
\arguments{
\item{X}{matrix with one row for each trial}

\item{baseLength:}{n of baseline samples (counted from the left)}
}
\description{
Normalize epochs with respect to baseline (e.g. useful for pupil size)
}
