% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-1.R
\name{bisquare_loss}
\alias{bisquare_loss}
\title{Robust loss functions}
\usage{
bisquare_loss(err, k)
}
\description{
Tukey's bisquare loss function. (quadratic at first, then tapers off).
Note that you need to have an estimate of the residual's standard deviation (call it sigma_hat)
to set the parameter k, which determines where the loss function is quadratic and where it is not
typical values for k, which provide high efficiency also in the normal case, are then 
k = sigma_hat * 4.685
}
