% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych-6.R, R/psych-7.R
\name{boot_3par_multi}
\alias{boot_3par_multi}
\title{Psychometric}
\usage{
boot_3par_multi(p, d, l_b, u_b, nsim = 1000)

boot_3par_multi(p, d, l_b, u_b, nsim = 1000)
}
\arguments{
\item{p}{parameter vector, c(mu, sigma, lambda)}

\item{d}{a dataframe with a column `r` that indicate response (0,1) and a column `x` that indicate the stimulus, plus a column `condition`}

\item{p}{parameter vector, c(mu, sigma, lambda)}

\item{d}{a dataframe with a column `r` that indicate response (0,1) and a column `x` that indicate the stimulus}
}
\description{
Bootstrap model to estimate parameter posterior distribution (useful for calculating confidence intervals). Multiple conditions and fixed lapse rate.

Bootstrap model to estimate parameter posterior distribution (useful for calculating confidence intervals)
}
