% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootfooci.R
\name{bootFooCI}
\alias{bootFooCI}
\title{Miscellaneous helpers}
\usage{
bootFooCI(v, nsim = 1000, FUN = mean(), method = "bca",
  alpha = 0.05, ...)
}
\arguments{
\item{nsim}{number of bootstrap simulations (default is 1000)}

\item{FUN}{is the function that compute the statistic of interest}

\item{method}{Either "bca" (default, accellerated & bias-corrected, Efron, 1987), or "percentile"}

\item{alpha}{Either "bca" (accellerated & bias-corrected, Efron, 1987), or "percentile"}
}
\value{
the 95% confidence limits
}
\description{
Bootstrapped 95% CI for arbitratry function foo (BCA / percentile method)
}
