% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-display-2.R
\name{deg2pix}
\alias{deg2pix}
\title{Display measures}
\usage{
deg2pix(xRes, screenWidth, viewDistance, DEGS = 1)
}
\arguments{
\item{xRes}{horizontal screen resolution}

\item{screenWidth}{screen width in cm}

\item{viewDistance}{viewing distance in cm}

\item{DEGS}{how many degree to convert?}
}
\value{
pixels
}
\description{
Convert pixels to degrees of visual angles
}
