% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-6.R
\name{fillGap}
\alias{fillGap}
\title{Signal processing}
\usage{
fillGap(x, sp = 1/60, max = 50, type = "linear")
}
\arguments{
\item{x}{signal}

\item{sp}{sampling period}

\item{max}{maximum gape length (to be interpolated)}
}
\description{
gap fill-in  (cubic-spline/linear interpolation)
I used that to interpolate blink in pupil size recordings (e.g., Lisi Bonato & Zorzi, Biologial Psychology, 2015)
}
