% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-1.R
\name{filtLP}
\alias{filtLP}
\title{Signal processing}
\usage{
filtLP(x, fc, sp = 1/60)
}
\arguments{
\item{x}{signal}

\item{fc}{cut-off frequency}

\item{sp}{sampling period}
}
\description{
ifinite-impulse-response (IIR) single-pole lowpass filter
}
