% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix-circular-5.R
\name{fit_mixture}
\alias{fit_mixture}
\title{Circular-mixture}
\usage{
fit_mixture(v, se = T, method = "noboot", nsim = 1000)
}
\arguments{
\item{par}{vector: p(guess), mean direction, k}

\item{d}{vector: vector of direction angles}
}
\description{
wrapper function to fit the model and output the parameters as list
}
