% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych-2.R
\name{lnorm_3par}
\alias{lnorm_3par}
\title{Psychometric}
\usage{
lnorm_3par(p, d)
}
\arguments{
\item{p}{parameter vector, c(mu, sigma, lambda)}

\item{d}{a dataframe with a column `r` that indicate response (0,1) and a column `x` that indicate the stimulus, plus a column `condition`}
}
\description{
simple (negative) log-likelihood calculation for 1 condition
}
