% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix-circular-3.R
\name{negLogLik_mix}
\alias{negLogLik_mix}
\title{Circular-mixture}
\usage{
negLogLik_mix(par, d)
}
\arguments{
\item{par}{vector: p(guess), mean direction, k}

\item{d}{vector: vector of direction angles}
}
\description{
compute the negative log-likelihood for the mixture model
}
