% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-1.R
\name{outfilter}
\alias{outfilter}
\title{General}
\usage{
outfilter(rtv, sv, nsd = 2)
}
\arguments{
\item{rtv}{varable of interest (vector)}

\item{sv}{grouping factor}

\item{nsd}{distance from the mean at which is applied the cut-off (in standard deviations)}
}
\description{
Adaptive outlier cut-off
}
