% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-display.R
\name{pix2deg}
\alias{pix2deg}
\title{Display measures}
\usage{
pix2deg(xRes, screenWidth, viewDistance, PIXEL = 1)
}
\arguments{
\item{xRes}{horizontal screen resolution}

\item{screenWidth}{screen width in cm}

\item{viewDistance}{viewing distance in cm}
}
\value{
degrees
}
\description{
Convert pixels to degrees of visual angles
}
