% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych-4.R
\name{predict_3par_multi}
\alias{predict_3par_multi}
\title{Psychometric}
\usage{
predict_3par_multi(p, d)
}
\arguments{
\item{p}{parameter vector, c(mu, sigma, lambda)}

\item{d}{a dataframe with a column `r` that indicate response (0,1) and a column `x` that indicate the stimulus, plus a column `condition`}
}
\description{
predicted choice probability (multiple conditions, constant lapse rate)
}
