% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savage_dickey_bf.R
\name{savage.dickey.bf}
\alias{savage.dickey.bf}
\title{Savage-Dickey density ratio}
\usage{
savage.dickey.bf(x, x_0 = 0, prior.mean = 0, prior.sd = 1,
  plot = F)
}
\arguments{
\item{x}{vector of samples from posterior distribution}

\item{x_0}{location of point-null hypothesis (usally zero)}

\item{prior.mean}{mean of Gaussian prior used to estiamte the model}

\item{prior.sd}{standard deviation of Gaussian prior used to estiamte the model}

\item{plot}{logical: should the prior and posterior probability be plotted?}
}
\value{
Bayes factor, indicating how likely are the data under the null hypothesis, compared to the alternative.
}
\description{
Approximate the Savage-Dickey density ratio, and compute Bayes factor in favour of null hypothesis.
As input it takes a vector of samples, which should represent the posterior distribution of the quantity of interests, usually obtained through MCMC sampling.
The function assumes a Gaussian prior (the default is mean 0 and sd 1).
}
