% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych-5.R
\name{simulate_3par_multi}
\alias{simulate_3par_multi}
\title{Psychometric}
\usage{
simulate_3par_multi(p, d)
}
\arguments{
\item{p}{parameter vector, c(mu, sigma, lambda)}

\item{d}{a dataframe with a column `r` that indicate response (0,1) and a column `x` that indicate the stimulus, plus a column `condition`}
}
\description{
Simulate psy function (e.g. for bootstrapping)
}
