% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal-7.R
\name{splitGazeData}
\alias{splitGazeData}
\title{Signal processing}
\usage{
splitGazeData(x, evento, flag, tw, sp = 1/60, n_before = 0.5,
  maxEpochNa = 0.1)
}
\arguments{
\item{x}{signal}

\item{evento}{factor with lenght equal to x (signal) indicating for each sample the trial event ("fixation", "target", etc.}

\item{flag}{selected event to align with (epochs will be time locked to the beginning of this event)}

\item{tw}{length of temporla window in sec}

\item{sp}{sampling period (secs, e.g. 1/60)}

\item{maxEpochNa}{max proportion of missing values (NA) in retained epochs (0:1)}

\item{flag}{cut-off frequency}
}
\description{
Align different epochs based on event of interest (useful for Tobii Eyetrackers)
}
